Rd cambios

+++ src/api/routes.py
@@ -11,6 +11,8 @@ from slugify import slugify
 api = Blueprint('api', __name__)
 
 ## REVISAR PARA BORRAR TODAS LAS FUNCIONES QUE VENIAN EN EL boilerplate ##
+
+
 @api.route('/hello', methods=['POST', 'GET'])
 def handle_hello():
     response_body = {
@@ -28,12 +30,13 @@ def handle_hello():
 def register():
     email = request.json.get("email", None)
     password = request.json.get("password", None)
+
     if email is None:
         return jsonify({"msg": "Email is missing, write one, please!"}), 404
     if password is None:
         return jsonify({"msg": "Password is missing, write one, please!"}), 404
 
-    user = User.query.filter_by(email = email).first()
+    user = User.query.filter_by(email=email).first()
 
     if user is not None:
         return jsonify({"msg": "Sorry. User already exist!"}), 409
@@ -48,6 +51,7 @@ def register():
 ################################################################################
 #                                   Login                                      #
 ################################################################################
+
 
 @api.route("/login", methods=["POST"])
 def login():
@@ -64,8 +68,9 @@ def login():
     return jsonify({"token": access_token, "user_id": user.id}), 200
 
 ################################################################################
-#                            CRUD de Post                                      #
+#                            CRUD de group                                      #
 ################################################################################
+
 
 @api.route("/group", methods=["POST"])
 @jwt_required()
@@ -105,14 +110,17 @@ def post_group():
 def update_group(id):
     data = request.get_json()
     owner_id = get_jwt_identity()
-    ##Hacer un if para comprobar que el owner_id sea el mismo que me viene en el token
+    keys = list(data.keys())
     group = Group.query.get(id)
-    if data["name"] is not None:
+
+    if owner_id != group.owner_id:
+        return jsonify({"msg": "you can´t upgrade this group"}), 404
+
+    if "name" in keys:
         group.name = data["name"]
-    if data["private"] is not None:
+    if "private" in keys:
         group.private = data["private"]
 
-    db.session.commit()
     return jsonify(group.serialize()), 200
 
 
@@ -132,6 +140,8 @@ def delete_group(id):
     owner_id = get_jwt_identity()
     if owner_id != group.owner_id:
         return jsonify({"msg": "you can´t delete this group"}), 404
+    if group is None:
+        return jsonify({"msg": "Group does not exist"}), 402
     db.session.delete(group)
     db.session.commit()
 
@@ -248,73 +258,77 @@ def update_event():
 #                           POST CRUD                                          #
 ################################################################################
 
+
 @api.route("/all_posts", methods=["GET"])
 @jwt_required()
 def get_all_post():
 
     post_array = []
-    
+
     all_post = Post.query.all()
 
     if len(all_post) == 0:
         return jsonify({"msg": "There is not post"}), 404
-    
-    for post in all_post:
-       post_array.append(post)
 
-    return jsonify([Post.serialize(post) for post in all_post]),200
+    for post in all_post:
+        post_array.append(post)
+
+    return jsonify([Post.serialize(post) for post in all_post]), 200
+
 
 @api.route("/all_user_posts", methods=["GET"])
 @jwt_required()
 def get_by_user():
     current_user_id = get_jwt_identity()
     post_array = []
-    all_post = Post.query.filter_by(user_id = current_user_id).all()
+    all_post = Post.query.filter_by(user_id=current_user_id).all()
     if len(all_post) == 0:
         return jsonify({"msg": "You don't have any posts"}), 404
-    
-    for post in all_post:
-       post_array.append(post)
 
-    return jsonify([Post.serialize(post) for post in all_post]),200
+    for post in all_post:
+        post_array.append(post)
+
+    return jsonify([Post.serialize(post) for post in all_post]), 200
 
 
 @api.route("/create_post", methods=["POST"])
 @jwt_required()
 def create_post():
-    text =  request.json.get("text", None)
-    image =  request.json.get("image", None)
+    text = request.json.get("text", None)
+    image = request.json.get("image", None)
     current_user_id = get_jwt_identity()
 
     new_post = Post(
-        text = text,
-        image = image,
-        user_id = current_user_id
+        text=text,
+        image=image,
+        user_id=current_user_id
     )
     db.session.add(new_post)
     db.session.commit()
-    return jsonify(Post.serialize(new_post)),200
+    return jsonify(Post.serialize(new_post)), 200
+
 
 @api.route("/update_post", methods=["PUT"])
 @jwt_required()
 def update_post():
     current_user_id = get_jwt_identity()
-    id =  request.json.get("id", None)
-    text =  request.json.get("text", None)
-    image =  request.json.get("image", None)
+    id = request.json.get("id", None)
+    text = request.json.get("text", None)
+    image = request.json.get("image", None)
 
-    post_exist = Post.query.filter_by(id = id).first()
+    post_exist = Post.query.filter_by(id=id).first()
 
     if post_exist is None:
-         return jsonify({"msg": "The post doesn't exist!"}), 404
+        return jsonify({"msg": "The post doesn't exist!"}), 404
 
     if text is not None:
-       post_exist.text = text
+        post_exist.text = text
     if image is not None:
         post_exist.image = image
 
     db.session.commit()
-    return jsonify(Post.serialize(post_exist)),200
+    return jsonify(Post.serialize(post_exist)), 200
+
 
 @api.route("/delete_post", methods=["DELETE"])
 @jwt_required()
@@ -325,14 +339,14 @@ def delete_post():
     if post_id is None:
         return jsonify({"msg": "Post ID is required!"}), 404
 
-    post_deleted = Post.query.filter_by(user_id = current_user_id).first()
+    post_deleted = Post.query.filter_by(user_id=current_user_id).first()
 
     if post_deleted is None:
         return jsonify({"msg": "The post is already deleted!"}), 404
 
     db.session.delete(post_deleted)
     db.session.commit()
-    return jsonify({"msg":"post deleted"}),200
+    return jsonify({"msg": "post deleted"}), 200
 
 ################################################################################
 #                           CRUD de User_Data                                  #
@@ -340,16 +354,14 @@ def delete_post():
 
 
 @api.route("/user/data/info", methods=["GET"])
-
 @jwt_required()
 def get_user_data():
     current_user_id = get_jwt_identity()
     current_user = User_Data.query.filter_by(user_id=current_user_id).first()
     if current_user is None:
 
-      
         return jsonify({"msg": "The user data does not exist"}), 400
-    return jsonify(current_user.serialize()),200
+    return jsonify(current_user.serialize()), 200
 
 
 @api.route("/user/data", methods=["POST"])
@@ -358,11 +370,11 @@ def post_user_data():
     data = request.get_json()
     current_user_id = get_jwt_identity()
     new_user_data = User_Data(
-        name = data["name"],
-        last_name = data["last_name"],
-        address = data["address"],
-        user_id = current_user_id,
-        profile_picture = None      #Por defecto dejo se crea sin profile_picture
+        name=data["name"],
+        last_name=data["last_name"],
+        address=data["address"],
+        user_id=current_user_id,
+        profile_picture=None  # Por defecto dejo se crea sin profile_picture
     )
     db.session.add(new_user_data)
     db.session.commit()
@@ -374,7 +386,7 @@ def post_user_data():
 def update_user_data():
     data = request.get_json()
     current_user_id = get_jwt_identity()
-    current_user = User_Data.query.filter_by(user_id = current_user_id).first()
+    current_user = User_Data.query.filter_by(user_id=current_user_id).first()
     keys = list(data.keys())
     if "name" in keys:
         current_user.name = data["name"]
@@ -409,27 +421,30 @@ def delete_user_data():
 @jwt_required()
 def get_all_image_user():
     current_user_id = get_jwt_identity()
-    images_user = Image.query.filter_by(owner_id = current_user_id).all()
+    images_user = Image.query.filter_by(owner_id=current_user_id).all()
     if len(images_user) == 0:
-        return jsonify({"msg": "this user has not images yet"}),400
+        return jsonify({"msg": "this user has not images yet"}), 400
     serializer = list(map(lambda picture: picture.serialize(), images_user))
-    return jsonify({"data": serializer}),200
+    return jsonify({"data": serializer}), 200
 
-## Estoy asumiendo que la image ya me viene en base64 y es lo que estoy guardando
+# Estoy asumiendo que la image ya me viene en base64 y es lo que estoy guardando
+
+
 @api.route("/user/image", methods=["POST"])
 @jwt_required()
 def post_image():
     current_user_id = get_jwt_identity()
     image = request.json.get("image", None)
     if image is None:
-        return jsonify({"msg": "no picture to upload"}),400
+        return jsonify({"msg": "no picture to upload"}), 400
     current_image = Image(
-        owner_id = current_user_id,
-        image = image
+        owner_id=current_user_id,
+        image=image
     )
     db.session.add(current_image)
     db.session.commit()
-    return jsonify({"msg": "the picture has been uploaded"}),200
+    return jsonify({"msg": "the picture has been uploaded"}), 200
+
 
 @api.route("/user/image/delete/<int:id>", methods=["DELETE"])
 @jwt_required()
@@ -439,4 +454,4 @@ def delete_image(id):
         return jsonify({"msg": "no picture to delete"})
     db.session.delete(image)
     db.session.commit()
-    return jsonify({"msg": "picture has been erased"}),200
+    return jsonify({"msg": "picture has been erased"}), 200
